Gracia Final support 3/6
git-svn-id: http://svn.l2jserver.com/trunk@2947 d34dfa0f-f91c-0410-86dd-cf06f19dd0ad

+++ L2_GameServer/java/net/sf/l2j/gameserver/model/itemcontainer/PetInventory.java
@@ -78,7 +78,7 @@ public class PetInventory extends Inventory
 		return (_items.size() + slots <= _owner.getInventoryLimit());
 	}
 	
-	public boolean validateWeight(L2ItemInstance item, int count)
+	public boolean validateWeight(L2ItemInstance item, long count)
 	{
 		int weight = 0;
 		L2Item template = ItemTable.getInstance().getTemplate(item.getItemId());
